pyproject.toml
[project]
name = "cc-genealogy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dask>=2025.1.0",
    "duckdb>=1.1.3",
    "huggingface-hub>=0.27.1",
    "pandas>=2.2.3",
    "pyarrow>=19.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.9.3",
    "tqdm>=4.67.1",
]
